Create a specific CCJob class and separate the tag counter example
